Learning my way though Laravel. After registration or login or anything else related to auth users will be redirected to main page.
