+++ routes/web.php
@@ -13,4 +13,4 @@
 
 Auth::routes();
 
-Route::get('/', 'HomeController@index')->name('home');
+Route::get('/', 'CommentController@index')->name("home");
